Time: 0 ms (100%), Space: 7.1 MB (18.45%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool judgeSquareSum(int c) {
         long long left = 0;
-        long long right = static_cast<long long>(sqrt(c));
+        long long right =sqrt(c);
         while (left <= right) {
             long long sum = left * left + right * right;
             if (sum == c) {
